refactor: drop proxies
