+++ src/components/App.vue
@@ -49,7 +49,7 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, onMounted, provide, ref, watch } from 'vue';
+import { computed, defineComponent, onMounted, ref } from 'vue';
 import { storeToRefs } from 'pinia';
 import { UrlParams } from '@vueuse/core';
 import vtkURLExtract from '@kitware/vtk.js/Common/Core/URLExtract';
@@ -64,12 +64,6 @@ import {
   loadUserPromptedFiles,
   loadUrls,
 } from '@/src/actions/loadUserFiles';
-import vtkResliceCursorWidget, {
-  ResliceCursorWidgetState,
-} from '@kitware/vtk.js/Widgets/Widgets3D/ResliceCursorWidget';
-import { useCurrentImage } from '@/src/composables/useCurrentImage';
-import type { Vector3 } from '@kitware/vtk.js/types';
-import { ViewTypes } from '@kitware/vtk.js/Widgets/Core/WidgetManager/Constants';
 import WelcomePage from '@/src/components/WelcomePage.vue';
 import { useDICOMStore } from '@/src/store/datasets-dicom';
 import LayoutGrid from '@/src/components/LayoutGrid.vue';
@@ -80,9 +74,7 @@ import KeyboardShortcuts from '@/src/components/KeyboardShortcuts.vue';
 import { useImageStore } from '@/src/store/datasets-images';
 import { useServerStore } from '@/src/store/server';
 import { useGlobalErrorHook } from '@/src/composables/useGlobalErrorHook';
-import { useWebGLWatchdog } from '@/src/composables/useWebGLWatchdog';
 import { useKeyboardShortcuts } from '@/src/composables/useKeyboardShortcuts';
-import { VTKResliceCursor } from '@/src/constants';
 
 export default defineComponent({
   name: 'App',
@@ -103,66 +95,7 @@ export default defineComponent({
     const dicomStore = useDICOMStore();
 
     useGlobalErrorHook();
-    useWebGLWatchdog();
     useKeyboardShortcuts();
-
-    const { currentImageData, currentImageMetadata } = useCurrentImage();
-
-    // --- ResliceCursorWidget --- //
-    // Construct the common instance of vtkResliceCursorWidget and provide it
-    // to all the child ObliqueView components.
-    const resliceCursor = vtkResliceCursorWidget.newInstance({
-      scaleInPixels: true,
-      rotationHandlePosition: 0.75,
-    }) as vtkResliceCursorWidget;
-    provide(VTKResliceCursor, resliceCursor);
-
-    // TODO: Move this to a store/global-state for reslicing.
-    // Orient the planes of the vtkResliceCursorWidget to the orientation
-    // of the currently set image.
-    const resliceCursorState =
-      resliceCursor.getWidgetState() as ResliceCursorWidgetState;
-
-    // Temporary fix to disable race between PanTool and ResliceCursorWidget
-    resliceCursorState.setScrollingMethod(-1);
-
-    watch(currentImageData, (image) => {
-      if (image && resliceCursor) {
-        resliceCursor.setImage(image);
-        // Reset to default plane values before transforming based on current image-data.
-        resliceCursorState.setPlanes({
-          [ViewTypes.YZ_PLANE]: {
-            normal: [1, 0, 0],
-            viewUp: [0, 0, 1],
-          },
-          [ViewTypes.XZ_PLANE]: {
-            normal: [0, -1, 0],
-            viewUp: [0, 0, 1],
-          },
-          [ViewTypes.XY_PLANE]: {
-            normal: [0, 0, -1],
-            viewUp: [0, -1, 0],
-          },
-        });
-        const planes = resliceCursorState.getPlanes();
-        if (currentImageMetadata.value) {
-          planes[ViewTypes.XY_PLANE].normal = currentImageMetadata.value
-            .lpsOrientation.Inferior as Vector3;
-          planes[ViewTypes.XY_PLANE].viewUp = currentImageMetadata.value
-            .lpsOrientation.Anterior as Vector3;
-
-          planes[ViewTypes.XZ_PLANE].normal = currentImageMetadata.value
-            .lpsOrientation.Anterior as Vector3;
-          planes[ViewTypes.XZ_PLANE].viewUp = currentImageMetadata.value
-            .lpsOrientation.Superior as Vector3;
-
-          planes[ViewTypes.YZ_PLANE].normal = currentImageMetadata.value
-            .lpsOrientation.Left as Vector3;
-          planes[ViewTypes.YZ_PLANE].viewUp = currentImageMetadata.value
-            .lpsOrientation.Superior as Vector3;
-        }
-      }
-    });
 
     // --- file handling --- //
