+++ src/components/tools/ResliceCursorTool.vue
@@ -1,174 +1,101 @@
-<script lang="ts">
-import { useViewStore } from '@/src/store/views';
-import {
-  computed,
-  defineComponent,
-  inject,
-  onBeforeUnmount,
-  ref,
-  toRefs,
-  watch,
-} from 'vue';
-import { useCurrentImage } from '@/src/composables/useCurrentImage';
-import vtkLPSView2DProxy from '@/src/vtk/LPSView2DProxy';
+<script setup lang="ts">
+import { computed, inject, ref, toRefs } from 'vue';
 import {
   vtkResliceCursorViewWidget,
   ResliceCursorWidgetState,
 } from '@kitware/vtk.js/Widgets/Widgets3D/ResliceCursorWidget';
-import { OBLIQUE_OUTLINE_COLORS, VTKResliceCursor } from '@/src/constants';
-import { onViewProxyMounted } from '@/src/composables/useViewProxy';
+import { OBLIQUE_OUTLINE_COLORS } from '@/src/constants';
 import { getLPSAxisFromDir, getVTKViewTypeFromLPSAxis } from '@/src/utils/lps';
 import { LPSAxisDir } from '@/src/types/lps';
-import { InitViewIDs } from '../../config';
+import useResliceCursorStore from '@/src/store/reslice-cursor';
+import { VtkViewContext } from '@/src/components/vtk/context';
+import { onViewMounted, onViewUnmounted } from '@/src/core/vtk/onViewMounted';
+import { InitViewIDs } from '@/src/config';
 
-export default defineComponent({
-  props: {
-    viewId: {
-      type: String,
-      required: true,
-    },
-    viewDirection: {
-      type: String,
-      required: true,
-    },
-  },
-  setup(props) {
-    const active = true;
-    const { viewId: viewID, viewDirection } = toRefs(props);
+interface Props {
+  viewId: string;
+  viewDirection: LPSAxisDir;
+}
 
-    const { currentImageMetadata: curImageMetadata, currentImageData } =
-      useCurrentImage();
+const props = defineProps<Props>();
+const { viewDirection } = toRefs(props);
+const viewAxis = computed(() => getLPSAxisFromDir(viewDirection.value));
 
-    const viewStore = useViewStore();
-    const viewProxy = computed(
-      () => viewStore.getViewProxy<vtkLPSView2DProxy>(viewID.value)!
-    );
-    const widgetManager = computed(() => viewProxy.value.getWidgetManager());
+const view = inject(VtkViewContext);
+if (!view) throw new Error('No vtk view');
 
-    const viewType = computed(() => {
-      return viewStore.viewSpecs[viewID.value].viewType;
+const resliceCursorStore = useResliceCursorStore();
+const { resliceCursor, resliceCursorState } = resliceCursorStore;
+
+const widget = ref<vtkResliceCursorViewWidget>();
+const vtkViewType = computed(() => getVTKViewTypeFromLPSAxis(viewAxis.value));
+
+onViewMounted(view.renderWindowView, () => {
+  widget.value = view.widgetManager.addWidget(
+    resliceCursor,
+    vtkViewType.value
+  ) as vtkResliceCursorViewWidget;
+
+  widget.value.setKeepOrthogonality(true);
+  // reset mouse cursor styles
+  widget.value.setCursorStyles({
+    translateCenter: 'default',
+    rotateLine: 'default',
+    translateAxis: 'default',
+  });
+
+  resliceCursorState.getStatesWithLabel('sphere').forEach((handle) => {
+    const h = handle as ResliceCursorWidgetState;
+    h.setScale1(10);
+    h.setOpacity(128);
+  });
+
+  resliceCursorState.getStatesWithLabel('line').forEach((handle) => {
+    const h = handle as ResliceCursorWidgetState;
+    h.setScale3(1, 1, 1);
+    h.setOpacity(100);
+  });
+
+  const xLines = [
+    ...resliceCursorState.getStatesWithLabel('XinZ'),
+    ...resliceCursorState.getStatesWithLabel('XinY'),
+  ];
+  xLines.forEach((handle) => {
+    const h = handle as ResliceCursorWidgetState;
+    h.setColor3(OBLIQUE_OUTLINE_COLORS[InitViewIDs.ObliqueSagittal]);
+  });
+
+  const yLines = [
+    ...resliceCursorState.getStatesWithLabel('YinZ'),
+    ...resliceCursorState.getStatesWithLabel('YinX'),
+  ];
+  yLines.forEach((handle) => {
+    const h = handle as ResliceCursorWidgetState;
+    h.setColor3(OBLIQUE_OUTLINE_COLORS[InitViewIDs.ObliqueCoronal]);
+  });
+
+  const zLines = [
+    ...resliceCursorState.getStatesWithLabel('ZinX'),
+    ...resliceCursorState.getStatesWithLabel('ZinY'),
+  ];
+  zLines.forEach((handle) => {
+    const h = handle as ResliceCursorWidgetState;
+    h.setColor3(OBLIQUE_OUTLINE_COLORS[InitViewIDs.ObliqueAxial]);
+  });
+
+  // update representation to not be as 3D
+  widget.value.getRepresentations().forEach((rep) => {
+    rep.getActors().forEach((actor) => {
+      actor.getProperty().setAmbient(1);
     });
+  });
 
-    const resliceCursor = inject(VTKResliceCursor);
-    if (!resliceCursor) {
-      throw new Error('Cannot fetch global instance of ResliceCursor');
-    }
+  view.requestRender();
+});
 
-    const state = resliceCursor.getWidgetState() as ResliceCursorWidgetState;
-    const widget = ref<vtkResliceCursorViewWidget>();
-
-    const VTKViewType = computed(() => {
-      const viewAxis = getLPSAxisFromDir(viewDirection.value as LPSAxisDir);
-      return getVTKViewTypeFromLPSAxis(viewAxis);
-    });
-
-    watch(widgetManager, (wm, oldWm) => {
-      if (oldWm) {
-        oldWm.removeWidget(resliceCursor);
-      }
-      if (wm) {
-        widget.value = wm.addWidget(
-          resliceCursor,
-          VTKViewType.value
-        ) as vtkResliceCursorViewWidget;
-      }
-    });
-
-    onViewProxyMounted(viewProxy, () => {
-      if (widgetManager?.value) {
-        widget.value = widgetManager.value.addWidget(
-          resliceCursor,
-          VTKViewType.value
-        ) as vtkResliceCursorViewWidget;
-
-        widget.value.setKeepOrthogonality(true);
-        // reset mouse cursor styles
-        widget.value.setCursorStyles({
-          translateCenter: 'default',
-          rotateLine: 'default',
-          translateAxis: 'default',
-        });
-
-        state.getStatesWithLabel('sphere').forEach((handle) => {
-          const h = handle as ResliceCursorWidgetState;
-          h.setScale1(10);
-          h.setOpacity(128);
-        });
-
-        state.getStatesWithLabel('line').forEach((handle) => {
-          const h = handle as ResliceCursorWidgetState;
-          h.setScale3(1, 1, 1);
-          h.setOpacity(100);
-        });
-
-        const xLines = [
-          ...state.getStatesWithLabel('XinZ'),
-          ...state.getStatesWithLabel('XinY'),
-        ];
-        xLines.forEach((handle) => {
-          const h = handle as ResliceCursorWidgetState;
-          h.setColor3(OBLIQUE_OUTLINE_COLORS[InitViewIDs.ObliqueSagittal]);
-        });
-
-        const yLines = [
-          ...state.getStatesWithLabel('YinZ'),
-          ...state.getStatesWithLabel('YinX'),
-        ];
-        yLines.forEach((handle) => {
-          const h = handle as ResliceCursorWidgetState;
-          h.setColor3(OBLIQUE_OUTLINE_COLORS[InitViewIDs.ObliqueCoronal]);
-        });
-
-        const zLines = [
-          ...state.getStatesWithLabel('ZinX'),
-          ...state.getStatesWithLabel('ZinY'),
-        ];
-        zLines.forEach((handle) => {
-          const h = handle as ResliceCursorWidgetState;
-          h.setColor3(OBLIQUE_OUTLINE_COLORS[InitViewIDs.ObliqueAxial]);
-        });
-
-        // update representation to not be as 3D
-        widget.value.getRepresentations().forEach((rep) => {
-          rep.getActors().forEach((actor) => {
-            actor.getProperty().setAmbient(1);
-          });
-        });
-
-        // show widget
-        viewProxy.value.renderLater();
-      }
-    });
-
-    onBeforeUnmount(() => {
-      if (widgetManager?.value) {
-        widgetManager.value.removeWidget(resliceCursor);
-      }
-    });
-
-    watch(
-      curImageMetadata,
-      (metadata) => {
-        state.placeWidget(metadata.worldBounds);
-      },
-      { immediate: true }
-    );
-
-    watch(
-      currentImageData,
-      (currImage) => {
-        if (resliceCursor && currImage) {
-          resliceCursor.setImage(currImage);
-        }
-      },
-      { immediate: true }
-    );
-
-    return {
-      active,
-      viewType,
-    };
-  },
+onViewUnmounted(view.renderWindowView, () => {
+  widget.value = undefined;
+  view.widgetManager.removeWidget(resliceCursor);
 });
 </script>
