+++ src/components/vtk/VtkLayerSliceRepresentation.vue
@@ -13,7 +13,7 @@ import { useSliceConfig } from '@/src/composables/useSliceConfig';
 import { LayerID, useLayersStore } from '@/src/store/datasets-layers';
 import useLayerColoringStore from '@/src/store/view-configs/layers';
 import { useLayerConfigInitializer } from '@/src/composables/useLayerConfigInitializer';
-import { applyColoring } from '@/src/composables/useColoringEffectNew';
+import { applyColoring } from '@/src/composables/useColoringEffect';
 
 interface Props {
   viewId: string;
