+++ src/composables/useColoringEffect.ts
@@ -1,102 +1,103 @@
-import { useProxyManager } from '@/src/composables/useProxyManager';
 import { Maybe } from '@/src/types';
 import {
-  ColorBy,
   ColorTransferFunction,
   ColoringConfig,
   OpacityFunction,
 } from '@/src/types/views';
-import vtkLPSView3DProxy from '@/src/vtk/LPSView3DProxy';
-import vtkVolumeRepresentationProxy from '@kitware/vtk.js/Proxy/Representations/VolumeRepresentationProxy';
-import { Mode as LookupTableProxyMode } from '@kitware/vtk.js/Proxy/Core/LookupTableProxy';
-import { Ref, computed, watchEffect } from 'vue';
+import { MaybeRef, computed, unref, watchEffect } from 'vue';
 import vtkPiecewiseFunctionProxy from '@kitware/vtk.js/Proxy/Core/PiecewiseFunctionProxy';
-import { getShiftedOpacityFromPreset } from '@/src/utils/vtk-helpers';
-import vtkAbstractRepresentationProxy from '@kitware/vtk.js/Proxy/Core/AbstractRepresentationProxy';
-import vtkProxyManager from '@kitware/vtk.js/Proxy/Core/ProxyManager';
+import {
+  applyNodesToPiecewiseFunction,
+  applyPointsToPiecewiseFunction,
+  getShiftedOpacityFromPreset,
+} from '@/src/utils/vtk-helpers';
+import vtkColorTransferFunction from '@kitware/vtk.js/Rendering/Core/ColorTransferFunction';
+import vtkColorMaps from '@kitware/vtk.js/Rendering/Core/ColorTransferFunction/ColorMaps';
+import vtkPiecewiseFunction from '@kitware/vtk.js/Common/DataModel/PiecewiseFunction';
+import vtkPiecewiseGaussianWidget from '@kitware/vtk.js/Interaction/Widgets/PiecewiseGaussianWidget';
 
 export interface ApplyColoringParams {
-  colorBy: ColorBy;
-  colorFunc: ColorTransferFunction;
-  opacityFunc: OpacityFunction;
-  rep: vtkAbstractRepresentationProxy;
-  proxyManager: vtkProxyManager;
+  props: {
+    colorFunction: ColorTransferFunction;
+    opacityFunction: OpacityFunction;
+  };
+  cfun: vtkColorTransferFunction;
+  ofun: vtkPiecewiseFunction;
+  componentIndex?: number;
 }
 
 export function applyColoring({
-  colorBy,
-  colorFunc,
-  opacityFunc,
-  rep,
-  proxyManager,
+  props: { colorFunction, opacityFunction },
+  cfun,
+  ofun,
+  componentIndex = -1,
 }: ApplyColoringParams) {
-  const { arrayName, location } = colorBy;
+  if (componentIndex === -1) {
+    cfun.setVectorModeToMagnitude();
+  } else {
+    cfun.setVectorModeToComponent();
+    cfun.setVectorComponent(componentIndex);
+  }
 
-  const lut = proxyManager.getLookupTable(arrayName);
-  lut.setMode(LookupTableProxyMode.Preset);
-  lut.setPresetName(colorFunc.preset);
-  lut.setDataRange(...colorFunc.mappingRange);
+  cfun.setMappingRange(...colorFunction.mappingRange);
+  const preset = vtkColorMaps.getPresetByName(colorFunction.preset);
+  if (preset) {
+    cfun.applyColorMap(preset);
+  }
 
-  const pwf = proxyManager.getPiecewiseFunction(arrayName);
-  pwf.setMode(opacityFunc.mode);
-  pwf.setDataRange(...opacityFunc.mappingRange);
+  const { mappingRange } = opacityFunction;
+  ofun.setRange(...opacityFunction.mappingRange);
 
-  switch (opacityFunc.mode) {
+  switch (opacityFunction.mode) {
     case vtkPiecewiseFunctionProxy.Mode.Gaussians:
-      pwf.setGaussians(opacityFunc.gaussians);
+      vtkPiecewiseGaussianWidget.applyGaussianToPiecewiseFunction(
+        opacityFunction.gaussians,
+        256,
+        opacityFunction.mappingRange,
+        ofun
+      );
       break;
     case vtkPiecewiseFunctionProxy.Mode.Points: {
       const opacityPoints = getShiftedOpacityFromPreset(
-        opacityFunc.preset,
-        opacityFunc.mappingRange,
-        opacityFunc.shift,
-        opacityFunc.shiftAlpha
+        opacityFunction.preset,
+        opacityFunction.mappingRange,
+        opacityFunction.shift,
+        opacityFunction.shiftAlpha
       );
       if (opacityPoints) {
-        pwf.setPoints(opacityPoints);
+        applyPointsToPiecewiseFunction(ofun, opacityPoints, mappingRange);
       }
       break;
     }
-    case vtkPiecewiseFunctionProxy.Mode.Nodes:
-      pwf.setNodes(opacityFunc.nodes);
+    case vtkPiecewiseFunctionProxy.Mode.Nodes: {
+      applyNodesToPiecewiseFunction(ofun, opacityFunction.nodes, mappingRange);
       break;
+    }
     default:
+      throw new Error('Invalid opacity function mode encountered');
   }
-
-  // control color range manually
-  rep.setRescaleOnColorBy(false);
-  rep.setColorBy(arrayName, location);
 }
 
 export function useColoringEffect(
-  config: Ref<Maybe<ColoringConfig>>,
-  imageRep: Ref<Maybe<vtkVolumeRepresentationProxy>>,
-  viewProxy: Ref<vtkLPSView3DProxy>
+  config: MaybeRef<Maybe<ColoringConfig>>,
+  cfun: MaybeRef<vtkColorTransferFunction>,
+  ofun: MaybeRef<vtkPiecewiseFunction>
 ) {
-  const colorBy = computed(() => config.value?.colorBy);
-  const colorTransferFunction = computed(() => config.value?.transferFunction);
-  const opacityFunction = computed(() => config.value?.opacityFunction);
-
-  const proxyManager = useProxyManager();
+  const colorTransferFunction = computed(() => unref(config)?.transferFunction);
+  const opacityFunction = computed(() => unref(config)?.opacityFunction);
 
   watchEffect(() => {
-    const rep = imageRep.value;
-    const colorBy_ = colorBy.value;
     const colorFunc = colorTransferFunction.value;
     const opacityFunc = opacityFunction.value;
-    if (!rep || !colorBy_ || !colorFunc || !opacityFunc || !proxyManager) {
-      return;
-    }
+    if (!colorFunc || !opacityFunc) return;
 
     applyColoring({
-      colorBy: colorBy_,
-      colorFunc,
-      opacityFunc,
-      rep,
-      proxyManager,
+      props: {
+        colorFunction: colorFunc,
+        opacityFunction: opacityFunc,
+      },
+      cfun: unref(cfun),
+      ofun: unref(ofun),
     });
-
-    // Need to trigger a render for when we are restoring from a state file
-    viewProxy.value.renderLater();
   });
 }
