--- src/composables/usePersistCameraConfigNew.ts
@@ -1,62 +1,0 @@
-import { MaybeRef, Ref, computed, unref } from 'vue';
-import { Maybe } from '@/src/types';
-import { CameraConfig } from '@/src/store/view-configs/types';
-import useViewCameraStore from '@/src/store/view-configs/camera';
-import vtkCamera from '@kitware/vtk.js/Rendering/Core/Camera';
-import { vtkFieldRef } from '@/src/core/vtk/vtkFieldRef';
-import { syncRef } from '@vueuse/core';
-import { guardedWritableRef } from '@/src/utils/guardedWritableRef';
-
-export function usePersistCameraConfig(
-  viewID: MaybeRef<string>,
-  dataID: MaybeRef<Maybe<string>>,
-  camera: MaybeRef<vtkCamera>
-) {
-  const viewCameraStore = useViewCameraStore();
-
-  type KeyType = keyof CameraConfig;
-  const keys: KeyType[] = [
-    'position',
-    'focalPoint',
-    'viewUp',
-    'parallelScale',
-    'directionOfProjection',
-  ];
-
-  const cameraRefs = keys.reduce(
-    (refs, key) => ({
-      ...refs,
-      [key]: guardedWritableRef(
-        vtkFieldRef(camera, key),
-        (incoming) => !!incoming
-      ),
-    }),
-    {} as Record<KeyType, Ref<any>>
-  );
-
-  const config = computed(() =>
-    viewCameraStore.getConfig(unref(viewID), unref(dataID))
-  );
-
-  const configRefs = keys.reduce(
-    (refs, key) => ({
-      ...refs,
-      [key]: computed({
-        get: () => config.value?.[key],
-        set: (v) => {
-          const viewIDVal = unref(viewID);
-          const dataIDVal = unref(dataID);
-          if (!viewIDVal || !dataIDVal) return;
-          viewCameraStore.updateConfig(viewIDVal, dataIDVal, {
-            [key]: v,
-          });
-        },
-      }),
-    }),
-    {} as Record<KeyType, Ref<any>>
-  );
-
-  keys.forEach((key) => {
-    syncRef(configRefs[key], cameraRefs[key]);
-  });
-}
