+++ src/core/provider.ts
@@ -1,5 +1,4 @@
 import { DICOMIO } from '../io/dicom';
-import ProxyWrapper from './proxies';
 import PaintTool from './tools/paint';
 
 /**
@@ -7,16 +6,13 @@ import PaintTool from './tools/paint';
  */
 export function CorePiniaProviderPlugin({
   paint,
-  proxies,
   dicomIO,
 }: {
   paint?: PaintTool;
-  proxies?: ProxyWrapper;
   dicomIO?: DICOMIO;
 } = {}) {
   const dependencies = {
     $paint: paint ?? new PaintTool(),
-    $proxies: proxies,
     $dicomIO: dicomIO ?? new DICOMIO(),
   };
   return () => dependencies;
