+++ src/global.d.ts
@@ -1,6 +1,5 @@
 import 'pinia';
 import type { Framework } from 'vuetify/types';
-import ProxyWrapper from './core/proxies';
 import PaintTool from './core/tools/paint';
 import { DICOMIO } from './io/dicom';
 
@@ -8,7 +7,6 @@ declare module 'pinia' {
   export interface PiniaCustomProperties {
     // from CorePiniaProviderPlugin
     $paint: PaintTool;
-    $proxies: ProxyWrapper;
     $dicomIO: DICOMIO;
   }
 }
