+++ src/main.js
@@ -9,7 +9,6 @@ import '@kitware/vtk.js/Rendering/OpenGL/Profiles/Glyph';
 import { createApp } from 'vue';
 import VueToast from 'vue-toastification';
 import { createPinia } from 'pinia';
-import vtkProxyManager from '@kitware/vtk.js/Proxy/Core/ProxyManager';
 import vtkMapper from '@kitware/vtk.js/Rendering/Core/Mapper';
 import { setPipelinesBaseUrl, setPipelineWorkerUrl } from '@itk-wasm/image-io';
 
@@ -19,13 +18,10 @@ import vuetify from './plugins/vuetify';
 import { DICOMIO } from './io/dicom';
 import { FILE_READERS } from './io';
 import { registerAllReaders } from './io/readers';
-import proxyConfiguration from './vtk/proxy';
 import { CorePiniaProviderPlugin } from './core/provider';
-import ProxyManagerWrapper from './core/proxies';
 import { patchExitPointerLock } from './utils/hacks';
 import { init as initErrorReporting } from './utils/errorReporting';
 import { StoreRegistry } from './plugins/storeRegistry';
-import { PROXY_MANAGER_WRAPPER_INJECT_KEY } from './composables/useProxyManager';
 
 // patches
 patchExitPointerLock();
@@ -38,10 +34,6 @@ vtkMapper.setResolveCoincidentTopologyLineOffsetParameters(-3, -3);
 
 registerAllReaders(FILE_READERS);
 
-const proxyManagerWrapper = new ProxyManagerWrapper(
-  vtkProxyManager.newInstance({ proxyConfiguration })
-);
-
 const dicomIO = new DICOMIO();
 dicomIO.initialize();
 
@@ -52,7 +44,6 @@ setPipelinesBaseUrl(itkConfig.imageIOUrl);
 const pinia = createPinia();
 pinia.use(
   CorePiniaProviderPlugin({
-    proxies: proxyManagerWrapper,
     dicomIO,
   })
 );
@@ -62,7 +53,6 @@ const app = createApp(App);
 
 initErrorReporting(app);
 
-app.provide(PROXY_MANAGER_WRAPPER_INJECT_KEY, proxyManagerWrapper);
 app.use(pinia);
 app.use(VueToast);
 app.use(vuetify);
