+++ src/store/datasets-layers.ts
@@ -102,9 +102,6 @@ export const useLayersStore = defineStore('layer', () => {
       image = vtkITKHelper.convertItkToVtkImage(itkImage);
     }
 
-    this.$proxies.addData(id, image);
-
-    // calling after adding data to proxy manager delays enabling deletion, thus avoiding error
     this.layerImages[id] = image;
   }
 
@@ -149,10 +146,6 @@ export const useLayersStore = defineStore('layer', () => {
     );
 
     delete this.layerImages[layerToDelete.id];
-
-    // May have errored creating data, so check before delete
-    if (this.$proxies.getData(layerToDelete.id))
-      this.$proxies.deleteData(layerToDelete.id);
   }
 
   function getLayers(key: DataSelection | null | undefined) {
