+++ src/store/segmentGroups.ts
@@ -143,8 +143,6 @@ export const useSegmentGroupStore = defineStore('segmentGroup', () => {
     orderByParent.value[metadata.parentImage] ??= [];
     orderByParent.value[metadata.parentImage].push(id);
 
-    this.$proxies.addData(id, labelmap);
-
     return id;
   }
 
@@ -402,7 +400,6 @@ export const useSegmentGroupStore = defineStore('segmentGroup', () => {
 
         const id = useIdStore().nextId();
         dataIndex[id] = labelmapImage;
-        this.$proxies.addData(id, labelmapImage);
         return id;
       })
     );
